workflows/github-workflow/scripts/requirements.txt
edgegrid_python==1.3.1
python-dotenv==1.0.1
Requests==2.32.0
